+++ src/main/java/duke_exception/DukeException.java
@@ -1,4 +1,4 @@
-package exception;
+package duke_exception;
 
 /**
  * An Exception class used to handle invalid commands where Duke is unable to
